refactor(pms): route wms stock policy reads through integration client
