+++ app/wms/stock/services/lot_service.py
@@ -8,8 +8,8 @@ from fastapi import HTTPException
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from app.integrations.pms.contracts import ItemPolicy
 from app.wms.stock.models.lot import Lot
-from app.pms.export.items.contracts.item_policy import ItemPolicy
 from app.wms.stock.services.lots import ensure_internal_lot_singleton, ensure_lot_full
