+++ app/wms/stock/services/lots.py
@@ -9,8 +9,8 @@ from typing import Optional
 from sqlalchemy import text
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.pms.export.items.contracts.item_policy import ItemPolicy
-from app.pms.export.items.services.item_read_service import ItemReadService
+from app.integrations.pms.contracts import ItemPolicy
+from app.integrations.pms.inprocess_client import InProcessPmsReadClient
 
 
 def normalize_lot_code(code: str | None) -> tuple[str, str]:
@@ -153,7 +153,7 @@ async def _load_item_policy(
     *,
     item_id: int,
 ) -> ItemPolicy:
-    policy = await ItemReadService(session).aget_policy_by_id(item_id=int(item_id))
+    policy = await InProcessPmsReadClient(session).get_item_policy(item_id=int(item_id))
     if policy is None:
         raise ValueError("item_not_found")
     return policy
